chore: create data generator with common logic for ConstantGenerator-like ruby tasks
The base generator contains the write thread logic, all the event
handling and the fabric methods as well. Classes that inherit from it
must implement the #values= assignment method and the #values method
